Issue 5102: Change the VCRT from a list of pointers to a list of global-ranks.
Previous memory usage for communicator rank lists:
  Global data: Array[size of job], 3 int each
  Per comm:    Array[size of comm], 1 void* each (pointers to the global data)
               Array[size of comm], 1 int each (worldranks for CCMI)

New  memory usage for communicator rank lists:
  Global data: None
  Per comm:    Array[size of comm], 1 int each

I the event that each node needed extra information, the global data
list was a structure with support for ref counting.  We only used a
single field, so it was overkill.  Now the comm contains a simple list
of included ranks, which happens to be in the perfect form to pass to
CCMI in the geometry.  Also, since the VCRT is shared between dup'ped
comms, the corresponding duped geometries will also share the same
list of ranks.