Issue 5265: Improve ad_bgl MPI_File_sync() scaling on GPFS.
If the underlying file system is GPFS or PVFS2, then only
one rank needs to fsync().  Define an ad_bgl specific flag
to indicate when "fsync aggregation" is enabled. When
enabled, MPI_File_sync() will barrier and have the
aggregator rank do the fsync and allreduce the results
to all ranks.