Get rid of MPID_SendRequest_create and a few MPIU_Object_set_ref calls.
MPID_SendRequest_create only existed to create a request starting with
2 references.  By removing the reference decrement from the
request_complete call, it now only needs a single reference.  This
reference is removed by the MPI top-level code, and the request is
then freed.

MPID_Request_release was being overused.  In many places it was being
used to release objects that used to have 2 references so that they
could be destroyed.  Since we no longer inflate the refcount, it is
generally not needed now.

I also found a couple locations where MPIU_Object_set_ref was used to
set the reference to 1, which was the default value.