Issue 5528: Correct the global-var name: s/dcmf_executable_name/MPID_Executable_name/
Also, correct the location of the definition.  Specifically, the
variable should _NOT_ be instantiated in a header file, but instead it
should be declared as an extern var in the header file and
instantiated in a .c file.

Also, we don't need to use malloc() & free() for this.  Actually, we
don't use malloc() & free() for _ANYTHING_ in MPICH2, because it
provides MPIU_Malloc() & MPIU_Free().  But beyond that, the variable
is always set to either argv[0] or "FORTRAN".  So I changed to code to
either point at one or the other instead of malloc()ing the storage
and copying.