Memregion leak in MPI glue-level rendezvous
I found a bug in rendezvous protocol (memory regions stay pinned on
RR/Linux and eventually run out if you don't destroy them).

I got pure DMA working over the weekend on IB, and this was one of the
bugs I discovered.  I think this is right, but you might want to
verify the patch.

Signed-off-by: Joe Ratterman <jratt@us.ibm.com>