Issue 6018: 2) Correct the type-sig of the "type" field in msginfo
Since the "type" field contains values from an enum, I declared it as
that enum type.  GCC handled this, but XLC crashes because it is a
signed type, which then gets sign-extended when reading, which gives
terrible values.  I don't know how to create an unsigned enum, so I
changed it back to uint16_t.

To find this, I added a bunch of debugging to the bad-path.  I am
leaving it.