Issue 6064: Re-Fix handling of persistent Bsends
We were seeing a number of IntelANL2006 failures with the previous Bsend_init fix.

MPICH2 keeps internal pointers to the Bsend requests (to do
garbage-collection on the bsend-buffer) AND returns pointers to the
user.  However, it doesn't change the ref-count, which should be 2.  I
now inflate the ref-count.  Then, when MPICH2 internally tests for
completion successfully, the count only drops to 1 and the child
request isn't freed.  When the user finally tests, THEN it becomes 0
and it is freed.

If the user tests first, I'm pretty sure that MPICH2 will eventually
call test to ensure that it can freed the bsend-buffer during detach.

I also added long comments to this change and the strange line in the
last commit.