Move "peerrank" out of the header and back into the request.
Because the current version of PAMI does not provide any information
about the sender process in the new-msg callback, I had to add it to
the header myself.  Initially, the easiest way was to move the
"peerrank" field from the request into the header.  However,
"peerrank" on a sending node should be the destination rank, while the
header needs to contain the sender rank (i.e. self).  The first
solution was to temporarily set the header field just before actually
sending, and them set it back afterward.  This worked, but didn't
really make sense.

This is (the first half of) the correct solution: put the "peerrank"
field in the request object where it belongs, and rename the header
field.  That will remain in use until PAMI is complete (#22).