Issue 5060: Set appropriate env var default; add new env var names
This commit updates several PAMI_* environment variables to add
a PAMID_* version of the variable that is documented for end users.
The previous PAMI_* variables will be eliminated in a future commit.

Some variables are renamed to be consistent with the other variables.
For example, PAMID_CONTEXT_MAX instead of PAMI_MAX_CONTEXTS to be
consistent with PAMID_CONTEXT_POST.

All environment variable synonyms are deleted except for 'well known'
environment variables such as PAMID_EAGER and PAMID_RZV. However,
mispellings and plurality synonyms are deleted for even these
environment variables. This means that "PAMID_RVZ" is no longer
examined during startup.

A new environment variable, PAMI_THREAD_MULTIPLE, is added to
set the following group of environment variables all at once since
these variables are typically all set or not set together.
- PAMID_CONTEXT_POST
- PAMID_CONTEXT_MAX
- PAMID_COMMTHREADS