Issue 7207: Fixes for collectives and datatypes.
Punt to mpich for now on the non SOW collectives. PAMI_COLLECTIVES_SELECTION=2
overrides this punting.

Fix for gatherv with MPI_IN_PLACE and not providing full parameters on all nodes

Change #define for displacement arrays and fix minor build issues with the change

Punt to mpich for the logical types and fixes for some of the 2-element types

Punt to mpich for allgatherv and other protocols

Check return codes from MPI DT to PAMI DT for types we don't support

Add trace macro support to mpido_scatter

Additional fix for gatherv with MPI_IN_PLACE