Issue 8238:Fix aggregator selection when any bridge isn't part of the communicator (for COMM_WORLD)
This is trickier than the old code had assumed.  When a bridge was not part of the communicator/job, it picked rank 0 and returned -- assuming every rank came to the same conclusion.

But some ranks' bridge may not be part of the communicator/job while other ranks' bridge is part of the communicator/job.  So you can't just pick rank 0 and assume it will work (rank 0 may have it's own bridge).   You need to pick one rank (arbitrarily) to use as the bridge AND communicate that (not just return).  I use minloc to pick the first rank without a bridge to act as bridge.  Then do the usual allgather/sort/bcast of bridge info.

This fix is really only for comm_world.  Other fixes are needed for subcomm's.