Issue 7967: Use the 'queue_immediate' hint to disable PAMI_Send_immediate overflow.
When 'queue immediate' is disabled the PAMI_Send_immediate() function
will return PAMI_EAGAIN if the send operation could not be completed
inline. The default behavior is to enable 'queue immediate' which will
cause PAMI to allocate memory, copy the source data, and queue the send
immediate request for later processing.

This default behavior is unfortunate for mpich because the mpi request
will be completed as soon as the PAMI_Send_immediate() function returns
when, in fact, the data has not been sent at all. This can lead to
situations where mpich will unknowingly perform a send flood, such as
when using MPI[C]_Sendrecv in a ring algorithm, and result in excessive
memory allocation and can even cause the process to run out of memory.