Issue 7967: Disallow PAMI_Send_immediate() for internal mpich collectives.
This change takes advantage of the 'context_offset' parameter of the
MPID_Isend() function. This parameter is purely internal ito mpich and is
_not_ part of the external MPI_Isend() function signature. MPICH uses this
parameter to differentiate normal user point-to-point communication and
internal collective point-to-point communication. With this change when the
MPID_Isend() is invoked with the mpich 'context_offset' set to non-zero the
pamid code will use a version of SendMsg that does not consider the short
cutoff and therefore will only perform eager or rendezvous communication.

It is important to disallow PAMI_Send_immediate() at scale because the
default behavior of PAMI_Send_immediate() is to enable 'queue immediate'
which will cause PAMI to allocate memory, copy the source data, and queue the
send immediate request for later processing. This default behavior is
unfortunate for mpich because the mpi request will be completed as soon
as the PAMI_Send_immediate() function returns when, in fact, the data
has not been sent at all. This can lead to situations where mpich will
unknowingly perform a send flood, such as when using MPI[C]_Sendrecv in
a ring algorithm, and result in excessive memory allocation and can even
cause the process to run out of memory.