Issue 7967: Add support for 'zero byte' rendezvous sends.
Previously if the 'short' and 'eager' protocols were disabled, as when
the "limit" environment variables are set to zero, the rendezvous send
path would, in certain configurations, attempt to create a zero-length
memory region and assert. Similarly, some internal MPICH2 collectives
implemented over point-to-point attemt to do a zero-byte send from a
NULL send buffer which will result in an assert when the memory region
is created.

The solution is to separate the code flow for 'zero-byte' and 'normal'
rendezvous sends. This separation keeps the impact of the zero-byte send
support to the normal rendezvous send flow to a minimum. The zero-byte
code path is quite simple as the rget of the source buffer is not needed
and the only action is to send the rendezvous ack when the matching
receive is posted.

Signed-off-by: Su Huang <suhuang@us.ibm.com>