Add 'PAMID_PT2PT_LIMITS' env var to specify *all* point-to-point limit overrides
The entire point-to-point limit set is determined by three boolean
configuration values:
- 'is non-local limit'   vs 'is local limit'
- 'is eager limit'       vs 'is immediate limit'
- 'is application limit' vs 'is internal limit'

The point-to-point configuration limit values are specified in order and
are delimited by ':' characters. If a value is not specified for a given
configuration then the limit is not changed. All eight configuration
values are not required to be specified, although in order to set the
last (eighth) configuration value the previous seven configurations must
be listed. The 'k', 'K', 'm', and 'M' multipliers may be specified. For
example:

   PAMID_PT2PT_LIMITS=":::::::10k"

The configuration entries can be described as:
   0 - remote eager     application limit
   1 - local  eager     application limit
   2 - remote immediate application limit
   3 - local  immediate application limit
   4 - remote eager     internal    limit
   5 - local  eager     internal    limit
   6 - remote immediate internal    limit
   7 - local  immediate internal    limit

Examples:

   "10K"
     - sets the application internode eager (the "normal" eager limit)

   "10240::64"
     - sets the application internode eager and immediate limits

   "::::0:0:0:0"
     - disables 'eager' and 'immediate' for all internal point-to-point

Signed-off-by: Su Huang <suhuang@us.ibm.com>