Issue 8879: fix bug in the 'is_local_task' extension.
The extension was reporting a non-zero value when the specified task is
local, but the src/mpid/pamid/src/pt2pt/mpidi_sendmsg.c is expecting a
boolean value to be used in the point-to-point eager limit lookup.

On bgq, the reported value was '64' which caused the eager limit lookup
to go off the end of the array.

The pe 'shift value' is 0 which will result in the expression
(1UL << 0) and should be optimized out by the compiler.

Signed-off-by: Charles Archer <archerc@us.ibm.com>