D187003:  MPI_File_get_position_share fail in mpich2
Variable could be passed into io routines uninitialized.
A file read was supposed to write the current values into the uninit
variable, but under certain circumstances the read would be 0
bytes, leading to the read call returning 0 and leaving the value
uninitialized.

Simple fix is to always init to zero. This preserves the original
meaning of a 0 byte read.

Signed-off-by: Michael Blocksome <blocksom@us.ibm.com>