CPS 92BQ9Q: Allow f90 mod files to be installed in an arbitrary directory.
Instead of using 'includedir' to install the mod file headers the
automake 'modincdir' is used instead.  The normal configure processing
already checks for a user-specified 'modincdir' variable and, if not set,
defaults to use 'includedir'.

Prior to this change, if the 'modincdir' variable was set during
configure, as in the following ..

  ./configure --includedir=$some_dir modincdir=$some_dir/gfortran

.. then the mod include directory would correctly be used to search for
the mod files in the mpif90 script, however the mod files themselves
would not have been installed in that directory.